requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cobble==0.1.4
colorama==0.4.6
cssselect2==0.7.0
docx2pdf==0.1.8
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.1
h11==0.14.0
html5lib==1.1
htmldocx==0.0.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
mammoth==1.8.0
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python-headless==4.10.0.84
outcome==1.3.0.post0
packaging==24.1
pdf2docx==0.5.8
pdfkit==1.0.0
pillow==10.4.0
plum-dispatch==1.7.4
pycparser==2.22
pydyf==0.11.0
pyhtml2pdf==0.0.7
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyphen==0.16.0
PySocks==1.7.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
Spire.Pdf==10.7.1
termcolor==2.4.0
tinycss2==1.3.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
zopfli==0.2.3
flasgger>=0.9.7.1